# api/openapi.yaml
openapi: 3.0.3
info:
  title: Walwatch Auto-Renewal Keeper API
  description: API for managing blob auto-renewal on Walrus/Sui.
  version: 0.1.0
  contact:
    name: Walwatch
    url: https://walwatch.io
servers:
  - url: https://api.walwatch.io/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    OrgIdHeader:
      name: X-Org-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Organization UUID for scoping requests
...
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AlertEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        alertRuleId:
          type: string
          format: uuid
          nullable: true
        eventType:
          type: string
        severity:
          type: string
        message:
          type: string
        status:
          type: string
        firedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true

    # ── Analytics ──
    AnalyticsOverview:
      type: object
      properties:
        totalBlobs:
          type: integer
        activeBlobs:
          type: integer
        totalProjects:
          type: integer
        totalPolicies:
          type: integer
        totalWallets:
          type: integer
    AnalyticsStorage:
      type: object
      properties:
        totalSizeBytes:
          type: integer
        averageSizeBytes:
          type: number
        blobCount:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
              totalSize:
                type: integer
    AnalyticsCostResponse:
      type: object
      properties:
        totalCost:
          type: number
        byPeriod:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              cost:
                type: number
    AnalyticsForecastResponse:
      type: object
      properties:
        projectedCost:
          type: number
        byPeriod:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              projectedCost:
                type: number
    RenewalCountResponse:
      type: object
      properties:
        totalRenewals:
          type: integer

    # ── Audit Log ──
...
    RenewalJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        blobRegistrationId:
          type: string
          format: uuid
        policyId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        attempt:
          type: integer
        maxAttempts:
          type: integer
        lastError:
          type: string
          nullable: true
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Team ──
...
    AdminTriggerScanInput:
      type: object
      required:
        - justification
      properties:
        justification:
          type: string
        orgId:
          type: string
          format: uuid
    AdminRetryJobInput:
      type: object
      required:
        - justification
      properties:
        justification:
          type: string
    AdminTenantResponse:
      type: object
      properties:
        organization:
          $ref: '#/components/schemas/Organization'
        stats:
          type: object
          properties:
            totalJobs:
              type: integer
            totalBlobs:
              type: integer
            activeSchedules:
              type: integer
            totalAlerts:
              type: integer
    QueueStatus:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer

    # ── Feature Flags ──
...
paths:
...
  /analytics/overview:
    get:
      tags: [Analytics]
      summary: Get analytics overview
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
  /analytics/storage:
    get:
      tags: [Analytics]
      summary: Get storage analytics
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Storage analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStorage'
  /analytics/renewals:
    get:
      tags: [Analytics]
      summary: Get renewal count
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Renewal count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenewalCountResponse'
  /analytics/costs:
    get:
      tags: [Analytics]
      summary: Get cost analytics
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Cost analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsCostResponse'
  /analytics/forecasts:
    get:
      tags: [Analytics]
      summary: Get cost forecast
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Cost forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsForecastResponse'

  # ── Audit Logs ──
...
  /alert-events:
    get:
      tags: [Alert Events]
      summary: List alert events
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of alert events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertEvent'
  /alert-events/{id}/acknowledge:
    post:
      tags: [Alert Events]
      summary: Acknowledge alert event
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert event acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertEvent'

  # ── Schedules ──
...
  /admin/queues:
    get:
      tags: [Admin]
      summary: Get queue status
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Queue statuses
          content:
            application/json:
              schema:
                type: object
                properties:
                  queues:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueueStatus'
  /admin/trigger-scan:
    post:
      tags: [Admin]
      summary: Trigger scan
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTriggerScanInput'
      responses:
        '200':
          description: Scan triggered
  /admin/retry-job/{id}:
    post:
      tags: [Admin]
      summary: Admin retry renewal job
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRetryJobInput'
      responses:
        '200':
          description: Job retry enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job:
                    $ref: '#/components/schemas/RenewalJob'
  /admin/tenants/{orgId}:
    get:
      tags: [Admin]
      summary: Get tenant details
      security:
        - AdminKeyAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTenantResponse'
  /admin/encryption/rotate-key:
    post:
      tags: [Admin]
      summary: Rotate encryption key
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - justification
              properties:
                justification:
                  type: string
      responses:
        '200':
          description: Key rotation result

  # ── Feature Flags ──
...